% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{phish_pal}
\alias{phish_pal}
\title{Function to interpolate a color palette}
\usage{
phish_pal(palette = "rift", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in phish_palettes}

\item{reverse}{Boolean true if palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\description{
Function to interpolate a color palette
}
