% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{phish_palettes}
\alias{phish_palettes}
\title{Color palettes}
\format{
An object of class \code{list} of length 8.
}
\usage{
phish_palettes
}
\description{
Color palettes
}
\keyword{datasets}
