% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{scale_color_phish}
\alias{scale_color_phish}
\title{Color scale for Phish colors}
\usage{
scale_color_phish(palette = "rift", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in phish_palettes}

\item{discrete}{Boolean if color aesthetic is discrete}

\item{reverse}{Boolean indicating whether palette should be reversed}

\item{...}{Additional arguments used to discrete_scale() or scale_fill_gradient()
to automatically interpolate between colours.}
}
\description{
Color scale for Phish colors
}
