% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{scale_fill_phish}
\alias{scale_fill_phish}
\title{Fill scale for Phish colors}
\usage{
scale_fill_phish(palette = "rift", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in phish_palettes}

\item{discrete}{Boolean if color aesthetic is discrete}

\item{reverse}{Boolean if palette should be reversed}

\item{...}{Additional arguments used to discrete_scale() or scale_fill_gradientn()
to automatically interpolate between colours.}
}
\description{
Fill scale for Phish colors
}
